requirements.txt
# Async / HTTP
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
aiohttp>=3.8.5        # updated for Python 3.12 compatibility
async-timeout==4.0.3
anyio==4.11.0
frozenlist==1.8.0
multidict==6.7.0
sniffio==1.3.1
yarl==1.22.0
websockets==15.0.1
httpcore==1.0.9
httpx>=0.27,<0.28
httpx-sse==0.4.3
httptools==0.7.1
h11==0.16.0

# Core framework
fastapi==0.120.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
uvicorn==0.38.0
uvloop==0.22.1
starlette==0.48.0
watchfiles==1.1.1
typer==0.20.0

# Supabase
supabase>=2.22.1
postgrest>=2.22.1       # updated for Python 3.12 compatibility
realtime>=2.22.1
storage3>=2.22.1
supabase-auth>=2.22.1
supabase-functions>=2.22.1

# OpenAI + LangChain
openai>=2.6.0
langchain>=0.3.79
langchain-community>=0.3.31
langchain-core>=0.3.79
langchain-openai>=0.3.35
langchain-text-splitters>=0.3.11
langsmith>=0.4.37
tiktoken==0.12.0

# Pydantic & type utilities
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.4
annotated-doc==0.0.2
annotated-types==0.7.0
typing-inspection==0.4.2
typing-inspect==0.9.0
typing_extensions==4.15.0

# Utilities
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
dataclasses-json==0.6.7
dnspython==2.7.0
distro==1.9.0
dotenv==0.9.9
python-dotenv==1.1.1
email-validator==2.3.0
exceptiongroup==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.0.2
orjson==3.11.3
packaging==25.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
shellingham==1.5.4
sentry-sdk==2.42.1
tenacity==9.1.2
tqdm==4.67.1
ujson==5.11.0
urllib3==2.5.0
zstandard==0.25.0

# PDF / OCR
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pypdfium2==4.30.0
pytesseract==0.3.13
pypdf
PyYAML==6.0.3
Pygments==2.19.2

# SQL / DB
SQLAlchemy==2.0.44

# Langchain / AI dependencies (already included above)
faiss-cpu==1.12.0
